chore: add more types to extension-api Signed-off-by: Sonia Sandler <ssandler@redhat.com>
